+++ HISTORY.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add a set of commands to interact with the cache
 - Show the author of a PR on the create screen
 - Make the team labels clickable in the create screen
+- Do not show PRs that have a ignored label in the create screen
 
 ## 0.4.0 - 2023-06-21